+++ HISTORY.rst
@@ -29,6 +29,9 @@ latest (unreleased)
 
 **Build**
 
+* The 'release.bump' task adds the entry in 'migration.yml' if it does not
+  exist
+
 **Documentation**